Make RunScript.Execute a private static method
There's not much of a need to instantiate a new RunScript object every
time we want to call a script so static solves that.

Secondarily, being able to run arbitrary scripts seems a bit not great?
By making `Execute` private, scripts will have to have their own method
defined in order to be invoked.