Remove unused Kumara.Tests project
Currently our structure is to have a separate test project for each
service in the repo leaving Kumara.Tests unused. We may want to reintroduce Kumara.Tests later when
we have a need for it but for now we can reduce our build and test time
slightly be removing this project.