Replace hardcoded dev credentials with generated credentials
A new make-secret script generates a secret using OpenSSL and
caches it in a directory within the repository git metadata.
This will give a stable output for a particular key within
the same checkout without requiring external housekeeping.

mise has its own caching too which we use for performance
reasons only. We can't rely on this for secret stability.