Always generate JSON Schema on Event Source startup
The current "already generated" test seems to not catch
all scenarios and it's very cheap to generate these schema
so the easy thing to do is just generate the files every time.

If we keep the POCO files as the source of truth, it's expected
we'll probably drop these files at some point and dynamically
generate the JSON when servicing metadata requests.