Add support for setting JSON property names at runtime
Attributes such as `[JsonPropertyName(…)]` only supports constant
expressions as parameters so this cannot be used to change the name
of a JSON attribute.

It is possible to use a JsonConverter to change how an object is
serialised to JSON but this has two main disadvantages:

  1. It is very low level and essentially requires taking over the
     entire serialisation process.

  2. Schema generation tools such as Swagger refuse to generate
     schema for types using JsonConverter as there is no way
     to tell what structure will be outputted.

Instead we can use a custom JSON type info resolver to make changes
to the type information .NET generates automatically via reflection.
This lets us focus on the changes we want to to make (in this case,
renaming some properties), without having to deal with everything.

Schema tools (e.g. Swagger) will still see the original JSON type info
so we have to make the same changes to that schema too. Not a biggie.