Add ITwinPathProvider which gets an iTwin's path to the root Account
iTwin Service provides a GetiTwinAncestors endpoint which seems
like it should do this in a single request but the Swagger states it
requires a subClass parameter (e.g. Account, Portfolio, Project) and
is thus not useful as a generic way to get a list of ancestor iTwin IDs.
I tried calling it without the parameter and it doesn't error but it
also does not return any results.

One might also think that GetiTwinParent is a better endpoint to call
than GetiTwinById with a $select filter, but this is not wrapped by
the Bentley.ConnectCoreLibs libraries and does not appear to be used
much or at all in wider Bentley. GetiTwinById actually does two get
calls underneath (get self then get parent) to return detail about
the parent, which we don't actually care about here so GetiTwinById
is more efficient anyway. Hooray!