Stop Respawner from erroring when the database has no migrations yet
Respawner has a check which throws an exception if it is not
cleaning any tables. This is generally a good thing as it likely
indicates a configuration issue. On a new project though that
results in an error we don't want. This change detects the migration
history table existing, meaning we're probably wired up correctly,
without any other tables, and cleans the (presumably empty) history
table instead to avoid the error being raised.