Apply custom property names to JsonPropertyInfo
Microsoft.AspNetCore.OpenApi.OpenApiSchemaService (an internal class)
applies schema transfomers by iterating over `jsonTypeInfo.Properties`
then probing the JSON schema to see if said property exists before
recursing and invoking the transformer on descendant nodes.

We need to keep schema.Properties and jsonTypeInfo.Properties in sync
otherwise the traversal stops and nested properties are not transformed.

We could try and ensure JsonPropertyNamesSchemaPatcher is always the
last transformer so this is not an issue but it's safer to just make
sure everything's consistent. Tests will fail if the field name changes.

I tried delaying the Properties mutations until the end by using a
IOpenApiDocumentTransformer instead of a IOpenApiSchemaTransformer
but that resulted in weirdness like duplicate entries in the resulting
Components.Schema dictionary which I couldn't work out. This is simpler.