Revert "Don't track queries used in Query objects"
This reverts commit cff046de5f6af69736a26b442d4a476c04145617.

No tracking causes an issue with the Query tests as they're not quite
equivalent, the result of the query returns entities without a way to
identify their associations so any ShouldBeEquivalentTo assertions fail.
There's likely a way to solve this but I'm not sure what the best way
would be so I'm opting to revert this change for now.