Add citus to docker-compose.local.yml
Citus populates the docker-entrypoint-initdb with a SQL file to create
the citus extension. To avoid overwriting this file, we've scoped what
we're mounting to just the extra users script.

We also cloned and customised the extra users script to create the citus
extension on all databases. Otherwise it was only created on the
kumara-admin database.