Add EventTypeRegistry to help with Event Type validations and lookups
The plan is to use EventTypeRegistry from a custom ValidationAttribute
class. The problem is that ValidationAttribute classes aren't
instantiated by the dependency injection container, and therefore we
can't use DI to access EventTypeRegistry.

That's why we went with the Lazy Singleton Pattern for this class, as
that allows us to use EventTypeRegistry from anywhere, without DI.