Add script to delete DNS search zones added by GlobalProtect VPN
VPN configuration currently injects an "online.local" search zone
which seems a bit strange since ".local" for mDNS. A bad side-effect
of this is DNS resolution can be very slow to the point of requests
timing out while mDNS is trying to find hosts that don't exist.
A simple fix for this to delete DNS search zones using this script.