Use persistent volumes to store data protection keys
Shared data storage is required for multiple containers
to be able to have shared keys for encrypting cookies.

Without this we'll get errors like the following:
> AuthenticationFailureException: Unable to unprotect the message.

On Azure Kubernetes we use the Azure File provider which supports
multiple nodes having read/write access to the shared network mount.

Locally with kind (Kubernetes in Docker), the default provider does not
support ReadWriteMany but that's fine when we only have a single node.