Add support for configuring local ingress with the .NET Core dev cert
This makes it a lot easier to test ingress locally with HTTPS.
This is handy when you want to test HTTPS things in in the browser.

One might think using socat as a proxy might be good enough but
that won't set the X-Forwarded- headers correctly.

You will probably want to still use socat to temporarily forward
the expected localhost HTTPS port for OIDC, etc. This script will
output a suitable command you can run after provisioning completes.